refactor as per code review
